Testing cache module
