+++ scripts/aws_lambda_functions/sentryPythonDeleteTestFunctions/lambda_function.py
@@ -1,12 +1,12 @@
 import boto3
-import sentry_sdk 
+import sentry_sdk
 
 
 monitor_slug = "python-sdk-aws-lambda-tests-cleanup"
 monitor_config = {
     "schedule": {
         "type": "crontab",
-        "value": "0 12 * * 0", # 12 o'clock on Sunday
+        "value": "0 12 * * 0",  # 12 o'clock on Sunday
     },
     "timezone": "UTC",
     "checkin_margin": 2,
@@ -24,7 +24,7 @@ def delete_lambda_functions(prefix="test_"):
     """
     client = boto3.client("lambda", region_name="us-east-1")
     functions_deleted = 0
-    
+
     functions_paginator = client.get_paginator("list_functions")
     for functions_page in functions_paginator.paginate():
         for func in functions_page["Functions"]:
@@ -39,17 +39,17 @@ def delete_lambda_functions(prefix="test_"):
                     print(f"Got exception: {ex}")
 
     return functions_deleted
-    
+
 
 def lambda_handler(event, context):
     functions_deleted = delete_lambda_functions()
-    
+
     sentry_sdk.metrics.gauge(
-        key="num_aws_functions_deleted", 
+        key="num_aws_functions_deleted",
         value=functions_deleted,
     )
-    
+
     return {
-        'statusCode': 200,
-        'body': f"{functions_deleted} AWS Lambda functions deleted successfully."
+        "statusCode": 200,
+        "body": f"{functions_deleted} AWS Lambda functions deleted successfully.",
     }
